+++ system/apps/project_boards/lib/project_boards/shape_lifecycle_to_board_shapes/shape_lifecycle_to_board_shapes.ex
@@ -63,7 +63,9 @@ defmodule ProjectBoards.ShapeLifecycleToBoardShapes.ShapeLifecycleToBoardShapes 
       points: field(:points, data),
       color: field(:color, data),
       width: field(:width, data),
-      text: field(:text, data)
+      text: field(:text, data),
+      from_shape_id: field(:from_shape_id, data),
+      to_shape_id: field(:to_shape_id, data)
     }
 
     if Store.new_shape?(shape.shape_id) do