+++ system/apps/query_boards/lib/query_boards/get_board_snapshot_by_id_over_mesh/get_board_snapshot_by_id_over_mesh.ex
@@ -133,7 +133,9 @@ defmodule QueryBoards.GetBoardSnapshotByIdOverMesh.GetBoardSnapshotByIdOverMesh 
       points: field(:points, shape),
       color: field(:color, shape),
       width: field(:width, shape),
-      text: field(:text, shape)
+      text: field(:text, shape),
+      from_shape_id: field(:from_shape_id, shape),
+      to_shape_id: field(:to_shape_id, shape)
     }
   end